[refactor] Fixed batch bugs, and file structure
Updated the following:
1. DynamicSemaphore now has its own file.
2. Parser logic now is inside the llm folder
3. batch_chat now uses aync_non_stream.
4. batch_chat now has a default of 5 retries per request.
