+++ llmstudio/__init__.py
@@ -1,2 +1,3 @@
 from llmstudio.llm import LLM
-from llmstudio.parser import Parser
+from llmstudio.llm.parser import Parser
+from llmstudio.llm.dynamic_semaphore import DynamicSemaphore
